docs: add content to readme
